+++ install-software-scripts/12-kde.sh
@@ -4,15 +4,15 @@ echo -e '\nConfiguring KDE\n'
 
 mkdir -p /tmp/kde/.config
 
-mv ~/.config/discoverrc /tmp/kde/.config
-mv ~/.config/dolphinrc /tmp/kde/.config
-mv ~/.config/kdeglobals /tmp/kde/.config
-mv ~/.config/kglobalshortcutsrc /tmp/kde/.config
-mv ~/.config/konsolerc /tmp/kde/.config
-mv ~/.config/konsolesshconfig /tmp/kde/.config
-mv ~/.config/plasmarc /tmp/kde/.config
-mv ~/.config/xsettingsd /tmp/kde/.config
-mv ~/.config/ksmserverrc /tmp/kde/.config
-mv ~/.local/share/konsole /tmp/kde
+if [ -f ~/.config/discoverrc ]; then mv ~/.config/discoverrc /tmp/kde/.config; fi
+if [ -f ~/.config/dolphinrc ]; then mv ~/.config/dolphinrc /tmp/kde/.config; fi
+if [ -f ~/.config/kdeglobals ]; then mv ~/.config/kdeglobals /tmp/kde/.config; fi
+if [ -f ~/.config/kglobalshortcutsrc ]; then mv ~/.config/kglobalshortcutsrc /tmp/kde/.config; fi
+if [ -f ~/.config/konsolerc ]; then mv ~/.config/konsolerc /tmp/kde/.config; fi
+if [ -f ~/.config/konsolesshconfig ]; then mv ~/.config/konsolesshconfig /tmp/kde/.config; fi
+if [ -f ~/.config/plasmarc ]; then mv ~/.config/plasmarc /tmp/kde/.config; fi
+if [ -d ~/.config/xsettingsd ]; then mv ~/.config/xsettingsd /tmp/kde/.config; fi
+if [ -f ~/.config/ksmserverrc ]; then mv ~/.config/ksmserverrc /tmp/kde/.config; fi
+if [ -d ~/.local/share/konsole ]; then mv ~/.local/share/konsole /tmp/kde; fi
 
 stow  -v -d ~/git/linux-setup/dotfiles/ -t ~ kde
