+++ install-software-scripts/70-keepass.sh
@@ -5,6 +5,6 @@ echo -e '\nInstalling KeePassXC\n'
 #KeepassXC
 echo -e 'Installing KeepassXC\n'
 sudo apt install keepassxc
-mv ~/.config/KeePass /tmp
+if [ -d KeePass ]; then mv ~/.config/KeePass /tmp; fi
 stow -v -d ~/git/linux-setup/dotfiles/ -t ~ keepass
